Get rid of the temp arena when parsing variables
Previously, the parser in id_str and ck_act used a "temp arena" to store token
values during lexing. The idea was that the token values would be freed after
parsing (maybe even after parsing a single line), but needed to live longer than
just a single token, as parsing a line needed data from several tokens.

By being more careful in ck_act to always MM_ArenaStrdup variable names (and
function names) immediately after lexing, we can ensure that only the most
recent token's data needs to be available. Then, we can just store the token as
part of the ParserState. (It may be needed later if we are peeking a token, but
we can only peek one token ahead, so this is fine.)

This nicely reduced the memory needed, and the complexity, but does require a
bit more thought in the ck_act parser.